fix: swapped page names
